[ARCTIC-341][CORE]support TracerReplacePartitions impl ReplacePartitions to commit file change info. (#344)
* resolve 341

* fix code style

* support sync pos-delete file info in UnkeyedTable

* add unit test

* fix spark delete cause empty snapshot

* fix spark delete cause empty snapshot

+++ spark/v3.1/spark/src/main/java/com/netease/arctic/spark/writer/UnkeyedSparkBatchWrite.java
@@ -173,17 +173,21 @@ public class UnkeyedSparkBatchWrite implements ArcticSparkWriteBuilder.ArcticWri
     @Override
     public void commit(WriterCommitMessage[] messages) {
       RowDelta rowDelta = table.newRowDelta();
-      for (DeleteFile file : WriteTaskDeleteFilesCommit.deleteFiles(messages)) {
-        rowDelta.addDeletes(file);
+      if (WriteTaskDeleteFilesCommit.deleteFiles(messages).iterator().hasNext()) {
+        for (DeleteFile file : WriteTaskDeleteFilesCommit.deleteFiles(messages)) {
+          rowDelta.addDeletes(file);
+        }
+        rowDelta.commit();
       }
-      rowDelta.commit();
 
 
       AppendFiles appendFiles = table.newAppend();
-      for (DataFile file : WriteTaskDeleteFilesCommit.dataFiles(messages)) {
-        appendFiles.appendFile(file);
+      if (WriteTaskDeleteFilesCommit.dataFiles(messages).iterator().hasNext()) {
+        for (DataFile file : WriteTaskDeleteFilesCommit.dataFiles(messages)) {
+          appendFiles.appendFile(file);
+        }
+        appendFiles.commit();
       }
-      appendFiles.commit();
     }
   }
